requirements.txt
pdf2image
pytesseract
gradio
transformers
torch
Pillow
PyMuPDF
